front-end adjustments

+++ routes/routes.php
@@ -1,25 +1,25 @@
 <?php
 
-    require_once('../model/Membro.class.php');
-    require_once('../model/Advertencia.class.php');
-    require_once('../controller/MembersController.class.php');
-    require_once('../controller/AdvertenciasController.class.php');
+    require_once("../model/Membro.class.php");
+    require_once("../model/Advertencia.class.php");
+    require_once("../controller/MembersController.class.php");
+    require_once("../controller/AdvertenciasController.class.php");
 
     session_start();
-    if(isset($_POST['loginAttempt'])){
-        $login = $_POST['login'];
-        $password = $_POST['password'];
+    if(isset($_POST["loginAttempt"])){
+        $login = $_POST["login"];
+        $password = $_POST["password"];
 
         $user = new Membro($login, $password);
         if($user->auth()){
-            $_SESSION['auth'] = true;
-            $_SESSION['login'] = $_POST["login"];
-            $_SESSION['role'] = $user->getRole();
-            $_SESSION['name'] = $user->getName();
-            $_SESSION['privilege'] = $user->getPrivilege();
-            $_SESSION['score'] = $user->getScore();
-            $_SESSION['id'] = $user->getId();
-            if($_SESSION['privilege']){
+            $_SESSION["auth"] = true;
+            $_SESSION["login"] = $_POST["login"];
+            $_SESSION["role"] = $user->getRole();
+            $_SESSION["name"] = $user->getName();
+            $_SESSION["privilege"] = $user->getPrivilege();
+            $_SESSION["score"] = $user->getScore();
+            $_SESSION["id"] = $user->getId();
+            if($_SESSION["privilege"]){
                 header("location:../view/pcd.php");
             }
             else{
@@ -31,7 +31,7 @@
         }
     }
 
-    if(isset($_POST['registerAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["registerAttempt"])){
 
         $membersController = new MembersController();
 
@@ -45,11 +45,11 @@
         $member = new Membro($login, $password, $id=null, $score, $role, $name, $privilege);
         $membersController->registerMemberDB($member);
                 
-        header("location:../view/painel.php?valid=true");
+        header("location:../view/pcd.php?validRegisterMember=true");
 
     }
 
-    if(isset($_POST['updateMemberAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["updateMemberAttempt"])){
 
         $membersController = new MembersController();
 
@@ -64,37 +64,37 @@
         $member = new Membro($login, $password, $id, $score, $role, $name, $privilege);
         $membersController->updateMemberDB($member);
                 
-        header("location:../view/update.php?valid=true");
+        header("location:../view/pcd.php?validUpdateMember=true");
 
     }
 
-    if(isset($_POST['deleteMemberAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["deleteMemberAttempt"])){
         $membersController = new MembersController();
 
-        $id = $_POST['id'];
+        $id = $_POST["id"];
         
         $membersController->deleteMemberDB($id);
 
-        header("location:../view/pcd.php");
+        header("location:../view/pcd.php?validDeleteMember=true");
     }
 
-    if(isset($_POST['warningAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["warningAttempt"])){
 
         $advController = new AdvertenciasController();
         
-        $idmember = $_POST['idmember'];
-        $date = $_POST['date'];
-        $reason = $_POST['reason'];
-        $score = $_POST['score'];
-        $responsible = $_POST['responsible'];
-        $dismissed = $_POST['dismissed'];
+        $idmember = $_POST["idmember"];
+        $date = $_POST["date"];
+        $reason = $_POST["reason"];
+        $score = $_POST["score"];
+        $responsible = $_POST["responsible"];
+        $dismissed = $_POST["dismissed"];
         
     
         $adv = new Advertencia($id=null, $date, $reason, $score, $responsible, $dismissed, $idmember);
         $advController->registerWarningDB($adv);
 
         if($adv->okay()){
-            $_SESSION['okay'] = true;
+            $_SESSION["okay"] = true;
             $mbmController = new MembersController();
             $member = $mbmController->getMemberDB($adv->getIdMember());
             $newScore = ($member[4] - $score);
@@ -103,41 +103,43 @@
             echo $member;
 
             $mbmController->updateMemberScore($idmember, $newScore);
-            header("location:../view/paineladv.php?send=true");
+            header("location:../view/pcd.php?validRegisterWarning=true");
         }
         
     }
     
-    if(isset($_POST['updateWarningAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["updateWarningAttempt"])){
 
         $wngController = new AdvertenciasController();
 
-        $id = $_POST['id'];
-        $idmember = $_POST['idmember'];
-        $date = $_POST['date'];
-        $reason = $_POST['reason'];
-        $score = $_POST['score'];
-        $responsible = $_POST['responsible'];
-        $dismissed = $_POST['dismissed'];
+        $id = $_POST["id"];
+        $idmember = $_POST["idmember"];
+        $date = $_POST["date"];
+        $reason = $_POST["reason"];
+        $score = $_POST["score"];
+        $responsible = $_POST["responsible"];
+        $dismissed = $_POST["dismissed"];
 
         $wng = new Advertencia($id, $date, $reason, $score, $responsible, $dismissed, $idmember);
         $wngController->updateWarningDB($wng);
                 
-        header("location:../view/updateWarning.php?valid=true");
+        header("location:../view/pcd.php?validUpdateWarning=true");
 
     }
 
-    if(isset($_POST['deleteWarningAttempt'])){
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["deleteWarningAttempt"])){
         $wngController = new AdvertenciasController();
 
-        $id = $_POST['id'];
+        $id = $_POST["id"];
         
         $wngController->deleteWarningDB($id);
 
-        header("location:../view/pcd.php");
+        header("location:../view/pcd.php?validDeleteWarning=true");
     }
 
-    if(isset($_POST['logoutAttempt'])) { 
+    if(isset($_SESSION["auth"]) && $_SESSION["auth"] = true && isset($_POST["logoutAttempt"])) { 
+        unset($_SESSION["auth"]);
+        unset($_SESSION["login"]);
         session_destroy(); 
         header("location:../index.php");
     }
